+++ code/misc/movement.cc
@@ -1598,7 +1598,7 @@ int TBeing::displayMove(dirTypeT dir, int was_in, int total)
 // was_in should equal -1 if you don't want player being returned to original rm
 // gets called recursively for all riders, so initial call should be for
 // lowest level of riding
-int TBeing::genericMovedIntoRoom(TRoom *rp, int was_in, 
+int TBeing::genericMovedIntoRoom(TRoom *rp, long was_in, 
      checkFallingT checkFall)
 {
   TThing *t, *t2, *t3=NULL;
@@ -1663,7 +1663,7 @@ int TBeing::genericMovedIntoRoom(TRoom *rp, int was_in,
       }
     }
     char buf[32];
-    sprintf(buf, "%d", was_in);
+    sprintf(buf, "%li", was_in);
     rc = tmons->checkResponses(this, 0, buf, CMD_RESP_ROOM_ENTER);
     if (IS_SET_DELETE(rc, DELETE_THIS)) {
       delete tmons;
